Move Lookup model from model.commercial to commercial.model.capi
