+++ commercial/app/model/capi/CapiAgent.scala
@@ -3,7 +3,6 @@ package commercial.model.capi
 import common.{AkkaAgent, Logging}
 import contentapi.ContentApiClient
 import model.ContentType
-import model.commercial.Lookup
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
