+++ commercial/app/model/merchandise/events/MasterclassAgent.scala
@@ -1,12 +1,11 @@
 package commercial.model.merchandise.events
 
 import commercial.model.Segment
-import commercial.model.capi.Keyword
+import commercial.model.capi.{Keyword, Lookup}
 import commercial.model.feeds.{FeedMetaData, ParsedFeed}
+import commercial.model.merchandise.{Masterclass, MerchandiseAgent}
 import common.{ExecutionContexts, Logging}
 import contentapi.ContentApiClient
-import commercial.model.merchandise.{Masterclass, MerchandiseAgent}
-import model.commercial.Lookup
 
 import scala.concurrent.Future
